CORE-9238 Refactor EditMetadataTemplate components into their own files.
