+++ react-components/src/metadata/admin/SlideUpTransition.js
@@ -1,0 +1,7 @@
+/**
+ * @author psarando
+ */
+import React from "react";
+import Slide from "@material-ui/core/Slide";
+
+export default (props) => (<Slide direction="up" {...props} />);
